feat(S11P12A804-253): elasticsearch 설치 및 설정, 옷 제목 검색 기능 구현

See merge request s11-webmobile1-sub2/S11P12A804!31

+++ backend/clothes/src/main/java/com/ssafy/ozz/clothes/category/repository/CategoryLowRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.ozz.clothes.category.repository;
 import com.ssafy.ozz.clothes.category.domain.CategoryLow;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategoryLowRepository extends JpaRepository<CategoryLow, Long> {
+public interface CategoryLowRepository extends JpaRepository<CategoryLow, Byte> {
 }
